Re-facto EntryConsumer
Using an abstract method allow to share code but also can be used it we add a new broker in the future
